structs-methods-interfaces/shapes.go
package smi

import "math"

type Rectangle struct {
	Width  float64
	Height float64
}

type Circle struct {
	Radius float64
}

type Triangle struct {
	Width  float64
	Height float64
}

type Shape interface {
	Area() float64
}

func Perimeter(rt Rectangle) float64 {
	return 2 * (rt.Height + rt.Width)
}

func (r Rectangle) Area() float64 {
	return r.Height * r.Width
}

func (c Circle) Area() float64 {
	return math.Pow(c.Radius, 2) * math.Pi
}

func (t Triangle) Area() float64 {
	return t.Height * t.Width / 2
}
